Disable flow for shopping-event-conversion
flow does not understand this piece of code that merges the two objects.

  const data : FptiInput =
        Object.assign(
                getDefaultFPTIEvent(),
                  { ...defaultConvertedEvent,
                       eventData:
                            JSON.stringify(defaultConvertedEvent.eventData)
                  }
            );

It wants the first object returned by getDefaultFPTIEvent() to act like
an FPTIInput with the mandatory event_type etc included. It does not
understand that Object.assign may take a partially consturcted object to
output a fully constructed object that adhers to the type definition.